Ran comparison of Scoring Methods
This folder contains data and results that demonstrate the intercorrelations between the various scoring methods for the AON shapes. These were run on all sample of English-language TED Talk transcripts, which variously do and do not conform to our normative narrative shape.

The key that you can use to understand these results:

Shape Scaling Method:
	lf = Linear Feature Scaling
	z = Z-scored

Shape Similarity Scoring Method:
	diff = Sequential Difference Scoring methods
	frech = Frechet distance scoring methods

Overall, there are differences in the resulting output between the various scaling and scoring methods. However, they seem to generally all converge on the same information. That is, the various "Narrativity" scores all intercorrelate around .50 and up and, in most cases, the intercorrelations are closer to .80+.